requirements.txt
streamlit
pandas
polars
plotly
numpy
scikit-learn
xgboost
simfin
python-dotenv
requests
